refactor: split MakiModel into separate models refactor: moved MakiTensor and MakiLayer into separate package
