+++ makiflow/core/base_layers.py
@@ -19,7 +19,7 @@ from abc import abstractmethod
 import tensorflow as tf
 from copy import copy
 import numpy as np
-from makiflow.core.maki_entities import MakiLayer, MakiTensor
+from makiflow.core.inference import MakiLayer, MakiTensor
 
 
 class BatchNormBaseLayer(MakiLayer):
