+++ makiflow/core/training/__init__.py
@@ -1,6 +1,7 @@
 from .athena import Athena
 from .hermes import Hermes
 from .tensorboard import TensorBoard
+from .loss_builder import Loss
 from abc import ABC
