+++ makiflow/models/regressor/training_modules/abs_loss.py
@@ -17,7 +17,7 @@
 
 import tensorflow as tf
 from .additional_losses import BasicTrainingModule
-from makiflow.core.loss_builder import Loss
+from makiflow.core.training.loss_builder import Loss
 from makiflow.models.common.utils import print_train_info, moving_average
 from makiflow.models.common.utils import new_optimizer_used, loss_is_built
 from tqdm import tqdm
