+++ makiflow/models/regressor/training_modules/additional_losses/compile.py
@@ -16,7 +16,7 @@
 # along with Foobar.  If not, see <https://www.gnu.org/licenses/>.
 
 from .perceptual_loss import PerceptualLossModuleGenerator
-from makiflow.core.maki_entities import MakiTensor
+from makiflow.core.inference import MakiTensor
 
 # Class that connects all additional losses
