+++ makiflow/models/ssd/ssd_model.py
@@ -20,10 +20,10 @@ from __future__ import absolute_import
 from makiflow.generators.pipeline.gen_base import GenLayer
 from makiflow.generators.ssd import SSDIterator
 from makiflow.layers import InputLayer, ConcatLayer, ActivationLayer
-from makiflow.core.maki_entities import MakiCore
+from makiflow.core.inference import MakiCore
 from makiflow.models.ssd.training_literals import TL
 from makiflow.models.ssd.ssd_utils import bboxes_wh2xy, bboxes_xy2wh
-from makiflow.core.loss_builder import Loss
+from makiflow.core.training.loss_builder import Loss
 
 import numpy as np
 import tensorflow as tf
