+++ makiflow/save_recover/builder.py
@@ -22,12 +22,12 @@ from makiflow.models.classificator import Classificator
 from makiflow.models.classificator.main_modules import CParams
 
 from makiflow.layers.trainable_layers import TrainableLayerAddress
-from makiflow.layers.untrainable_layers import UnTrainableLayerAddress, InputLayer
+from makiflow.layers.untrainable_layers import UnTrainableLayerAddress
 from makiflow.layers.rnn_layers import RNNLayerAddress
 
-from makiflow.core.maki_entities.maki_layer import MakiRestorable
-from makiflow.core.maki_entities.maki_tensor import MakiTensor
-from makiflow.core.maki_entities.maki_model import MakiModel
+from makiflow.core.graph_entities.maki_layer import MakiRestorable
+from makiflow.core.graph_entities.maki_tensor import MakiTensor
+from makiflow.core.inference.maki_model import MakiModel
 
 from makiflow.models.ssd.detector_classifier import DetectorClassifier, DCParams
 from makiflow.models import SSDModel
